+++ changelog.md
@@ -7,6 +7,7 @@
   - fixed timeout errors of waitSynchronized for RPC connections and accounts with high reliability
   - implemented region support for historical market data requests and socket connections
   - improved price and equity tracking
+  - added SL and TP fields to MetatraderDeal
 
 17.0.0
   - breaking change: updated typescript types